api/requirements.txt
# made. API — Runtime Dependencies
# Install: pip install -r api/requirements.txt

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
supabase>=2.3.0
redis>=5.0.0
httpx>=0.26.0
python-multipart>=0.0.6
